% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lc.R
\name{addDefaultLayout}
\alias{addDefaultLayout}
\title{Add a default layout to the opened web page}
\usage{
addDefaultLayout(layoutName)
}
\arguments{
\item{layoutName}{Type of the layout. See 'Details' for more information.}
}
\description{
\code{addDefaultLayout} adds a layout that can be later used to arrange charts on the page (by
default each new chart is added to the bottom of the page).
}
\details{
Currently the only supported type
of a default layout is table with arbitrary number of rows and columns.
To use it set the layout argument to \code{tableMxN}, where \code{N} is the
number of rows and \code{M} is the number of columns. Each cell will get an ID that consists of
a letter (indicating the row) and a number (indicating the column) (e.g. \code{B3} is an ID of
the second row and third column).
}
\examples{
\donttest{openPage(useViewer = FALSE)
addDefaultLayout("table3x2")}

}
