% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lc.R
\name{chartEvent}
\alias{chartEvent}
\title{Trigger an event}
\usage{
chartEvent(d, chartId, layerId = "main", event, sessionId = .id, app = .app)
}
\arguments{
\item{d}{ID of an element that triggered the event. May be index of a point or line, vector or
row and column indices for a heatmap, value of an input block (please, check \code{\link{lc_input}}
for more details about values). Should be \code{NULL} for \code{mouseout} or \code{marked} events.
N.B. This function is called from the web page and therefore all element indices start from zero as it is
used in JavaScript.}

\item{layerId}{ID of the layer. You can get IDs of all charts and their layers with \code{\link{listCharts}}.}

\item{event}{Type of event. Must be one of \code{"click", "mouseover", "mouseout", "marked", "labelClickRow", "labelClickCol"}.}

\item{sessionId}{ID of the session (opened client page) that triggered the event. The default value uses a local session
variable. This must be a single sessionId. You can get a list of IDs of all currently active with the
method \code{getSessionIds} inherited from class \code{\link{App}}. Possible errors in evaluation of
this argument are ignored.}

\item{app}{Object of class \code{\link{LCApp}} for which the event was triggered. Note that this argument is here for
internal use and is default value is a variable, stored in each session. If you are not use wrapper functions,
it is preferred to call method \code{chartEvent} of a corresponding instance of class \code{\link{LCApp}}.}

\item{id}{ID of the chart.}
}
\description{
This function is called whenever user clicks, selects or hover over elements of a chart. In turn,
it calls a corresponding callback function, if any was specified by the user. This function
is meant to be used internally. However, an experienced user can still use it to customize app
behavior in some complicated cases. This function can also emulate events triggered by non-existing
elements.
}
\details{
This function is a wrapper around method \code{chartEvent} of class \code{\link{LCApp}}.
}
\examples{
\donttest{x <- rnorm(50)
lc_scatter(x = x, y = 2*x + rnorm(50, 0.1), on_click = function(d) print(d))
chartEvent(51, "Chart1", "Layer1", "click")}

}
