% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lc.R
\name{closePage}
\alias{closePage}
\title{Stop server}
\usage{
closePage()
}
\description{
Stops the server and closes all currently opened pages (if any). This function is a
wrapper of the \code{stopServer} method inherited by the \code{\link{LCApp}} class from the \code{\link[jrc]{App}} class.
}
\examples{
\dontrun{openPage(useViewer = FALSE)
closePage()}

}
