% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lc.R
\name{dat}
\alias{dat}
\title{Link data to the chart}
\usage{
dat(...)
}
\arguments{
\item{...}{List of name values pair to define the properties.}
}
\description{
\code{dat} allows to link variables from the current environment to chart's properties.
On every \code{\link{updateCharts}} call all the data, provided via the \code{dat} function,
will be automatically reevaluated and the chart will be changed accordingly. One can also
put properties outside of the \code{dat} function, if they are going to be constant.
}
\examples{
\donttest{lc_scatter(dat(x = rnorm(30)), y = rnorm(30))
#note that the Y values remain the same after each updateCharts call
updateCharts()}

}
