% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lc.R
\name{lc_input}
\alias{lc_input}
\title{Add input forms to the page}
\usage{
lc_input(data = list(), place = NULL, ..., id = NULL)
}
\description{
\code{lc_input} adds an input form. This function is an rlc wrapper for an
HTML \code{<input>} tag. Five types of input are supported: \code{"text", "range", "checkbox", "radio" and "button"}.
}
\examples{
\donttest{lc_input(type = "checkbox", labels = paste0("el", 1:5), on_click = function(value) print(value),
value = T)
lc_input(type = "radio", labels = paste0("el", 1:5), on_click = function(value) print(value),
         value = 1)
lc_input(type = "text", labels = paste0("el", 1:5), on_click = function(value) print(value),
         value = c("a", "b", "c", "e", "d"))
lc_input(type = "range", labels = paste0("el", 1:5), on_click = function(value) print(value),
         value = 10, max = c(10, 20, 30, 40, 50), step = c(0.5, 0.1, 1, 5, 25))
lc_input(type = "button", labels = paste0("el", 1:5), on_click = function(value) print(value))}

}
