% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lc.R
\name{listCharts}
\alias{listCharts}
\title{List all existing charts and layers}
\usage{
listCharts()
}
\description{
\code{listCharts} prints a list of IDs of all existing charts and layers.
}
\examples{
\donttest{noise <- rnorm(30)
x <- seq(-4, 4, length.out = 30)

lc_scatter(dat(x = x,
               y = sin(x) + noise,
               colourValue = noise), 
           id = "plot", layerId = "points")
lc_line(dat(x = x, y = sin(x)), id = "plot", addLayer = TRUE)
lc_colourSlider(chart = "plot", layer = "points")

listCharts()}
}
