% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lc.R
\name{removeChart}
\alias{removeChart}
\title{Remove chart from the page}
\usage{
removeChart(id)
}
\arguments{
\item{id}{A vector of IDs of the charts to be removed.}
}
\description{
Removes an existing chart.
}
\examples{
\donttest{lc_scatter(dat(x = 1:10, y = 1:10 * 2), id = "scatter")
removeChart("scatter")}

}
