% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lc.R
\name{removeLayer}
\alias{removeLayer}
\title{Remove a layer from a chart}
\usage{
removeLayer(chartId, layerId)
}
\arguments{
\item{chartId}{ID of the chart from which to remove a layer.}

\item{layerId}{ID of the layer to remove.}
}
\description{
Removes a layer from an existing chart.
}
\examples{
\donttest{lc_scatter(dat(x = 1:10, y = 1:10 * 2), id = "scatter")
lc_abLine(a = 2, b = 0, id = "scatter", addLayer = TRUE)
removeLayer("scatter", "Layer1")}

}
