% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lc.R
\name{removeChart}
\alias{removeChart}
\title{Remove chart from the page}
\usage{
removeChart(chartId)
}
\arguments{
\item{chartId}{A vector of IDs of the charts to be removed.}
}
\description{
Removes an existing chart. Changes will be applied to all currently opened and future pages.
This function is a wrapper around method \code{removeChart} of
class \code{\link{LCApp}}.
}
\examples{
\donttest{lc_scatter(dat(x = 1:10, y = 1:10 * 2), chartId = "scatter")
removeChart("scatter")}

}
