% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lc.R
\name{removeLayer}
\alias{removeLayer}
\title{Remove a layer from a chart}
\usage{
removeLayer(chartId, layerId)
}
\arguments{
\item{chartId}{ID of the chart from which to remove a layer.}

\item{layerId}{ID of the layer to remove.}
}
\description{
Removes a layer from an existing chart. Changes will be applied to all currently opened and future pages.
This function is a wrapper around method \code{removeLayer} of
class \code{\link{LCApp}}.
}
\examples{
\dontrun{lc_scatter(dat(x = 1:10, y = 1:10 * 2), chartId = "scatter")
lc_abLine(a = 2, b = 0, chartId = "scatter", addLayer = TRUE)
removeLayer("scatter", "Layer1")}

}
